repeatable

+++ scripts/docker.sh
@@ -20,7 +20,14 @@ fi
 
 sudo apt update
 $APT_INSTALL apt-transport-https ca-certificates curl software-properties-common curl
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+
+if [[ ! -f "/usr/share/keyrings/docker-archive-keyring.gpg" ]]; then
+  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+fi
+
 echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
 
+sudo apt update
 $APT_INSTALL docker-ce
+
+sudo usermod -aG docker ${USER}
